[Refactor] PetService 통합
여러개의 PetService 하나로 통합
