--- src/main/java/ttakkeun/ttakkeun_server/service/PetService/PetCommandService.java
@@ -1,9 +1,0 @@
-package ttakkeun.ttakkeun_server.service.PetService;
-
-import ttakkeun.ttakkeun_server.dto.pet.PetRequestDTO;
-import ttakkeun.ttakkeun_server.entity.Member;
-import ttakkeun.ttakkeun_server.entity.Pet;
-
-public interface PetCommandService {
-    Pet add(PetRequestDTO.AddDTO request, Member member);
-}
