--- src/main/java/ttakkeun/ttakkeun_server/service/PetService/PetQueryService.java
@@ -1,8 +1,0 @@
-package ttakkeun.ttakkeun_server.service.PetService;
-
-import ttakkeun.ttakkeun_server.dto.pet.PetResponseDTO;
-import ttakkeun.ttakkeun_server.entity.Member;
-
-public interface PetQueryService {
-    PetResponseDTO.LoadResultDTO load(Long petId, Member member);
-}
